+++ packages/@expo/cli/src/start/server/type-generation/startTypescriptTypeGeneration.ts
@@ -60,6 +60,7 @@ export async function startTypescriptTypeGenerationAsync({
           projectRoot,
           getRouterDirectoryModuleIdWithManifest(projectRoot, exp)
         ),
+        plugin: exp?.extra?.router,
       }),
     ]);
   }