+++ packages/expo-router/plugin/src/index.ts
@@ -35,6 +35,8 @@ const withRouter: ConfigPlugin<
     asyncRoutes?: string | { android?: string; ios?: string; web?: string; default?: string };
     /** Should the sitemap be generated. Defaults to `true` */
     sitemap?: boolean;
+    /** Generate partial typed routes */
+    partialTypedGroups?: boolean;
   } | void
 > = (config, _props) => {
   const props = _props || {};