+++ packages/expo-router/src/link/useLinkHooks.ts
@@ -71,11 +71,9 @@ export type WebAnchorProps = {
 /**
  *
  */
-export interface LinkProps<T extends string | object>
-  extends Omit<TextProps, 'href'>,
-    WebAnchorProps {
+export interface LinkProps extends Omit<TextProps, 'href'>, WebAnchorProps {
   /** Path to route to. */
-  href: Href<T>;
+  href: Href;
 
   // TODO(EvanBacon): This may need to be extracted for React Native style support.
   /** Forward props to child component. Useful for custom buttons. */
@@ -126,9 +124,9 @@ export function useInteropClassName(props: { style?: TextProps['style']; classNa
 }
 
 export const useHrefAttrs = Platform.select<
-  (props: Partial<LinkProps<any>>) => { hrefAttrs?: any } & Partial<LinkProps<any>>
+  (props: Partial<LinkProps>) => { hrefAttrs?: any } & Partial<LinkProps>
 >({
-  web: function useHrefAttrs({ asChild, rel, target, download }: Partial<LinkProps<any>>) {
+  web: function useHrefAttrs({ asChild, rel, target, download }: Partial<LinkProps>) {
     return useMemo(() => {
       const hrefAttrs = {
         rel,