+++ docs/ui/components/TemplateBareMinimumDiffViewer/VersionSelector.tsx
@@ -1,5 +1,8 @@
 import { ChevronDownIcon } from '@expo/styleguide-icons/outline/ChevronDownIcon';
 
+import { betaVersion } from '~/package.json';
+const BETA_MAJOR_VERSION = betaVersion.split('.')[0];
+
 export type VersionSelectorProps = {
   version?: string | undefined;
   setVersion: (version: string) => void;
@@ -20,7 +23,9 @@ export const VersionSelector = ({
         onChange={e => setVersion(e.target.value)}>
         {availableVersions.map(version => (
           <option key={version} value={version}>
-            {version === 'unversioned' ? 'unversioned' : `SDK ${version}`}
+            {version === 'unversioned'
+              ? 'unversioned'
+              : `SDK ${version}${version === BETA_MAJOR_VERSION ? '  (beta)' : ''}`}
           </option>
         ))}
       </select>